Task-54567: Adapt design empty folder (#209)
Task-54567: Adapt design empty folder

Task-54567: Adapt design empty folder
