+++ documents-webapp/src/main/webapp/skin/less/documents.less
@@ -137,7 +137,7 @@
     .documents-body {
       min-height: calc(~"100vh - 300px");
       .documents-no-body {
-        height: 208px;
+        height: 276px;
       }
       .action-menu-item:hover {
         background-color: #e4e5e7;
